requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-slugify==8.0.4
